--- server/app/utils/firebase.py
@@ -1,4 +1,0 @@
-# app/utils/firebase.py
-def get_firebase_id_token():
-    # Implement token retrieval logic here if needed
-    pass
