+++ server/models.py
@@ -1,3 +1,4 @@
+# models.py
 import uuid
 from datetime import datetime
 from flask_sqlalchemy import SQLAlchemy
